make view works button its own component
